+++ src/ftxui/component/slider_test.cpp
@@ -2,8 +2,7 @@
 #include <stddef.h>       // for size_t
 #include <array>          // for array
 #include <ftxui/component/mouse.hpp>  // for Mouse, Mouse::Left, Mouse::Pressed, Mouse::Released
-#include <ftxui/dom/direction.hpp>  // for Direction, Direction::Down, Direction::Left, Direction::Right, Direction::Up
-#include <ftxui/dom/elements.hpp>   // for frame
+#include <ftxui/dom/elements.hpp>  // for GaugeDirection, GaugeDirection::Down, GaugeDirection::Left, GaugeDirection::Right, GaugeDirection::Up, frame
 #include <memory>  // for shared_ptr, __shared_ptr_access, allocator
 #include <string>  // for to_string
 
@@ -48,7 +47,7 @@ TEST(SliderTest, Right) {
       .min = 0,
       .max = 100,
       .increment = 10,
-      .direction = Direction::Right,
+      .direction = GaugeDirection::Right,
   });
   Screen screen(11, 1);
   Render(screen, slider->Render());
@@ -71,7 +70,7 @@ TEST(SliderTest, Left) {
       .min = 0,
       .max = 100,
       .increment = 10,
-      .direction = Direction::Left,
+      .direction = GaugeDirection::Left,
   });
   Screen screen(11, 1);
   Render(screen, slider->Render());
@@ -94,7 +93,7 @@ TEST(SliderTest, Down) {
       .min = 0,
       .max = 100,
       .increment = 10,
-      .direction = Direction::Down,
+      .direction = GaugeDirection::Down,
   });
   Screen screen(1, 11);
   Render(screen, slider->Render());
@@ -117,7 +116,7 @@ TEST(SliderTest, Up) {
       .min = 0,
       .max = 100,
       .increment = 10,
-      .direction = Direction::Up,
+      .direction = GaugeDirection::Up,
   });
   Screen screen(1, 11);
   Render(screen, slider->Render());